Started offsetting routes and controllers to seperate files to enhance modularity
