+++ server/middleware/upload.js
@@ -1,0 +1,8 @@
+import multer from "multer";
+
+// Configure multer storage (memory storage since files go to S3)
+const storage = multer.memoryStorage();
+
+const upload = multer({ storage: storage });
+
+export default upload;
